nbd: return list instead of generator for block status query
The NBD protocol says that the server will send exactly one structured
reply chunk for each selected metadata context. This means that the
block status data will be in a structured reply chunk, and will not be
broken into multiple chunks, so there is no need to return a generator
for the chunks.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>